basic testing of auth stuff
