+++ backend/package.json
@@ -12,7 +12,8 @@
     "body-parser": "^1.18.2",
     "express": "^4.16.3",
     "jsonwebtoken": "^8.2.0",
-    "mongoose": "^5.0.11"
+    "mongoose": "^5.0.11",
+    "supertest": "^3.0.0"
   },
   "devDependencies": {
     "eslint": "^4.19.1",
